Upgrade from Hadoop-BAM 5.3 to 6.0
This commit removes the checked in local maven repo and replaces
it with a remote private one. This helps to avoid checking large
files into this git repo.

In order to reduce dependencies, the BCF/VCF code has been removed
from Hadoop-BAM. I've talked with the Hadoop-BAM team about how
to publish maven artifacts that are compatible with all versions of
Hadoop. In the future, we will remove the Hadoop-BAM source
altogether and simply depend on a published artifact.