Split bam2adam command into 'bam2adam' and 'transform' commands
Previously, the 'bam2adam' command read BAM files using a Hadoop FileInputFormat,
performed one or more transforms and then save the transformed data to ADAM
format. This required the FileInputFormat to deal with the BGZF block compression
markers to find splits which was too error prone. We may revisit this in the
future.

Now, the 'bam2adam' command reads a BAM file locally using the SAMFileReader in
samtools and saves the ADAM/Parquet to the URI specified (e.g. hdfs://, s3://, local).
This command will be improved in the future batch entire directories of files
and support partitioning.

The new 'transform' command reads an ADAM/Parquet file, performs one or more
transformations in-memory and then saves the result to a new ADAM/Parquet file.

Additionally,
o Updated the version of samtools from 1.93 to 1.96-1534 per Chris Hartl's request
  to support his work on Base Quality Score Recalibration
o Dropped dependencies on Hadoop-BAM and Picard