Initial implementation of Base Quality Score Recalibration (BQSR)
This code was written and contributed by Chris Hartl as a
visiting researcher in the UC, Berkeley AMPLab while still an
employee at The Broad Institute. As such, this code contribution
is copyright The Broad Institute and licensed to the ADAM project
under the Apache 2.0 license.

The current version of GATK BQSR code is licensed to both the Broad and to
Appistry and cannot be included in ADAM; however, the original version of
BQSR, which only provides base qualities (BQSRv2 provides qualities
for cigar elements as well), was released open-source. This version
of BQSR is based on the unencumbered version of BQSR with the full
awareness and blessing of The Broad Institute.

This code should be considered an alpha release as it contains some
implementation bugs that need to be worked out (e.g. occasionally a
single node will be swamped with work and crash).