Missing brackets in VcfConverter.getType
There were missing brackets, I believe, in the method getType in the class VcfConverter,
which caused the VariantType.SV never to be returned correctly.

I've added a test to make sure that all the types are accessible to the getType method, and
I've moved the method itself to an object (since it depends on none of the fields of the
VcfConverter class itself).