Fix transform -sort_reads partitioning. Add -coalesce option to transform.
Previously, all unmappped reads when sorting had their referenceId set to
Int.MaxValue. This worked to place unmapped reads at the end of the file
but it also caused all unmapped reads to be placed in a single partition
causing shuffle skew. Unmapped reads are now placed in a range of positions
to distribute them more evenly.