Adding support for before, after clauses to SparkFunSuite.
I had a problem, with another branch, where I couldn't access the SparkContext inside a before{}
or after{} block in a SparkFunSuite.  This tries to fix that problem -- and refactors the creation
and destruction of the sparkcontext into separate methods for reuse.

I'm not sure if there's a problem though, with multi-threaded execution of tests?  This seems fragile,
somehow, and could use some comments/review.