ReferenceRegion represents a contiguous genomic region.
ReferenceRegion is a model representing a contiguous region of a single reference sequence.
I imagine using RR for:
- interval lists representing the target set in exome or gene panel sequencing
- a base for extending to support reference annotations
- binding events in ChIP-seq
- arbitrary "regions of interest" determined as an intermediate result during other calculations.

Anyway, we should have one base model supporting a lot of the common operations of interest
for regions in general.

Added support for Kryo serialization.

Indluded methods for calculating distance and overlap.