Added the spark.driver.port erasure to SparkFunSuite cleanup.
We've been getting intermittent errors, with respect to Spark being unable to bind to a port,
in the context of repeated unit tests.  This apparently is a known problem, see the thread here:
  http://blog.quantifind.com/posts/spark-unit-test/
and the follow-up from Matei here:
  https://groups.google.com/forum/#!topic/spark-users/MeVzgoJXm8I

The upshot is that we need to clearProperty('spark.driver.port') when we shut down our sparkContext
after a sparkTest.