Adding the adamRecordsRegionParquetLoad method and failing test.
This is an illustrative example only! AdamContext.adamRecordsRegionParquetLoad contains an
example of the kind of pushdown predicate we'd *like* to write, but can't (because of the
way that Parquet is written).  In the end, it seems, all the parameters to a pushdown
predicate need to be known at compile time.