Cosmetic changes to FastaConverter and FastaConverterSuite.
Just making a few little cosmetic changes to FastaConverter, to make part of the code more readable.

Also, added a second condition to one of the tests in FastaConverterSuite, to check that the reassembled
sequence (which had been created, but wasn't being used) matched what was in the original file.