Updating all instances of Seq[_] to Iterable[_] in RDD.groupBy()
Apparently, the signature for the RDD.groupBy method has changed (in 1.0) to return an Iterable
rather than a Seq.  This commit includes all the changes taht are needed to account for this change
downstream in our code, mostly updating types to Iterable and inserting a few calls to toSeq in cases
where that's not sufficient.