Adding RichADAMRecord.referenceLengthFromCigar
The new referenceLengthFromCigar method (and its associated ADAMRecord method, 'referenceLength')
is an implementation of a method which counts the length of an aligned read against the reference
sequence. This is useful because rec.start + rec.referenceLength is a logical "end" coordinate
to be used for testing overlap of an ADAMRecord against a reference range.