[ADAM-307] Fixing Features2ADAM so that it can be run from the CLI.
Features2ADAM fails because it doesn't include an 'index' field in its
second command-line argument, and therefore Args4j can't parse the Features2ADAMArgs
object off a command line.

This commit adds the missing 'index' argument to the Features2ADAMArgs class.

There is a test for this additional feature -- however, that test is marked as IGNORE
because we can't get it to work either locally or on Jenkins.  This should be addressed,
and I'll open up a ticket for it.  In the meantime, however, we've verified that manually
that this patch works as advertised.

Also modified PrintADAM to use the scala.Console.out rather than System.out.
This seems to allow us to (in tests) redirect the output stream using the
Console.withOut command.