Added table of (wall-clock) stage durations when the print_metrics option is used
The ADAMMetricsListener now records the wall-clock duration of each Spark stage. Previously only the total time taken for all of the tasks in the stage was recorded, but this is different from the wall-clock time on a multiple node cluster.

Also changed the text of the print_metrics option to clarify that metrics are printed to the log (as opposed to stdout)