[ADAM-476] Adding 'childLocators' to FileLocator
Updated the FileLocator trait to include a childLocators method, which
implements the ability to scan the children of a locator (in addition to
finding the parent).  This will be required for the AvroMultiParquetRDD
we're planning to introduce.  Unfortunately (see the comments) this
isn't possible to implement in HTTPFileLocator without additional
assumptions.

Also filled in the implementation of some missing methods in
HTTPFileLocator in this commit, and added comments to FileLocator
overall.