[ADAM-595] Delete the 'comparisons' code.
This means deleting the comparisons code here; these files (Comparisons,
AvailableComparisons, and GeneratorFilter) are moving into the qc-metrics
package, along with the ComparisonsTraversalEngine.

Also being deleted are the test-suites for each of these classes.

Finally, the CLI commands which use these classes, CompareADAM and
FindReads, are also being deleted.

Each of these classes will be moved into the qc-metrics repository as
part of a separate pull request.