Add Avro LRU cache for strings to reduce memory use
This commit adds a new Avro Binary Decoder that uses an LRU cache
of Utf8 (string) objects to minimize the creation of Uft8 objects.
If a Utf8 exists in the cache, its reference is used instead of
creating a new object. Moreover, the decoder readString() method
reuses a single Utf8 object when reading strings instead of
creating a new Utf8 object on each read.

The read() method in AvroSerializer was synchronized which isn't
necessary since Spark ensures that only a single thread will
call read. That sychronization is removed with this commit.