Splitting out SingleReadBucket into an Avro model
This is the beginning of the experiment we discussed prior to the Amplab
2016 retreat: can we split some of the model classes that are currently
hand-written into pairs:
(1) an Avro-schema'ed model class, and
(2) a "*Utils"-style class containing what were previously the class
methods of hte model.

The initial commit only converts the SingleReadBucket model, from a
Scala class into an Avro record, as an example of the work involved.

All the old class methods have been put into a new class,
SingleReadBucketUtils.

The old _object_ SingleReadBucket was shadowing (unavoidably) the
SingleReadBucket class, and so it has been renamed SingleReadBucketRDD.