[ADAM-940] Fix adam2vcf -sort_on_save flag
The -sort_on_save flag allows users to sort an adam file
before saving to VCF format.

The ADAMVCFOutputFormat stores VariantContextWritable objects
which are not Serializable. If a user requests a sort, this
elicits a shuffle which, prior to this commit, would exit
with an NPE at serialization.

The fix is to keep only convert to VariantContextWritable
objects immediately before saving since the Hadoop output
format understands how to write Writables.