[ADAM-1008] Modify jenkins-test script to support Java 8 build.
Resolves #1008. As per discussion with @shaneknapp, it is preferable to support
builds across Java versions by manually setting JAVA_HOME in a shell script run
inside of Jenkins. This is due to certain configuration limitations inside of
Jenkins. Currently, we are doing this via a script tracked inside of Jenkins.
I would rather roll all of that config into the ./scripts/jenkins-test script
so that we have revision control, etc.