[ADAM-1061] Fix attribute regex bug.
We had a bug in `org.bdgenomics.adam.util.AttributeUtils` where the regex for
splitting out the formatting string for array attributes was applied to all
attributes. In an array attribute (SAM "B" tags), the type of the array elements
is encoded before the attribute values, and is split off by commas. E.g.,
"B:i,1,2,3". If the attribute is a string (SAM "Z" tags), commas are allowed.
To resolve this, I split this regex into two regexes. We only apply the
regex for splitting out the array type if we are working on an array
attribute. This resolves #1061.