[ADAM-1233] Expose header lines in Variant-related GenomicRDDs
Resolves #1233. Adds a `headerLines` field to all Variant-related GenomicRDDs.
In the current implementation, this field is populated by the set of all
header lines for all INFO/FORMAT fields that we support. In a future patch, we
will also store the header lines that correspond to fields that do not map to
a specific field in the Variant/Genotype schema, and that are stored in an
attribute map.