[ADAM-1530] Support loading GO:query (S/CR/B)AMs as fragments.
Resolves #1530. Adds support for loading (S/CR/B)AM files that are globally
unsorted yet have group order set to query (read name) as Fragments.
Additionally, adds two unit tests. In the process of running the unit tests,
I found an issue in `SingleReadBucket` that was dropping a read from each
pair after the first pair in a partition due to an odd behavior in the
`Iterator.takeWhile` function, which drops the first item that does not
get taken.