[ADAM-1584] Add SortedGenomicRDD trait, support for sorted legacy files.
Resolves #1584, #1519, #1558. Adds a SortedGenomicRDD trait, and concrete
implementations of this trait for all datatypes. This trait eliminates the need
for the partition map (#1558), by tracking the underlying partitioner used to
partition the sorted data and an average record size for the sorted RDD.
This can be used to obtain bounds similar to the partition map, but can be
computed at a lower cost. Additionally, this data can be cheaply created by
peeking at the first item on each partition, which enables us to support legacy
formats including FASTA, coordinate sorted SAM/BAM/CRAM, and VCF where data is
sorted on disk but no partition map is available (#1584). When we infer the
partitioner, we can determine whether the partitioner is lexicographically
sorted or sorted by index by looking at the reference sequence order in the
partitioner and the sequence dictionary that is attached to the RDD. Since we
then use the inferred partitioner to co-partition data during a shuffle region
join, we can support both lexicographically and sequence index ordered sort
orders in the region join (#1519).