[ADAM-1611] Extend pipe APIs to Java, Python, and R.
Resolves #1611. Adds a Java-friendly pipe method onto `GenomicRDD`, and
conversion functions in the `org.bdgenomics.adam.api.java` package.
Simultaneously, we add a `GenomicRDD` base type in the R bindings. With
these additions, we then add the pipe method to the `GenomicRDD` class
that exists in both Python and R. The language bindings use reflection
to construct the in/out-formatters and conversion functions.