[ADAM-1614] Add VariantContextRDD to R and Python APIs.
Resolves #1614. Restructures both R and Python APIs to add `GenomicDataset`
class, which inherits from `GenomicRDD`, as in adam-core. Moves all implementing
classes onto `GenomicDataset`, and adds `VariantContextRDD` which is the lone
class that extends `GenomicRDD`. Deletes `saveAsVcf` from `GenotypeRDD` and
`VariantRDD`, and moves this functionality into `toVariantContextRDD` functions
and `VariantContextRDD.saveAsVcf`. Additionally, adds `sort` and
`sortLexicographically` to the R API, and fixes `sortLexicographically` in
the Python API (it was calling `sort` previously).