Changed many of the groupBy+Joins to Window
Because the Spark SQL implementation of duplciate marking
was not scaling well to cluster runs, this version convers
many of the groupBy followed by Join back to the original dataset
operations with Window functions. This hopefully reduces the
amount of data that has to be shuffled when running on a cluster
and will make the performance benefits scale.