Inverted duplicates are not found in mark duplicates
Mark duplicates uses the left and right positions corrected for clipping
to match duplicates on coordinates. However, this approach can miss
duplicates that are truly inverted (ie positions reversed and the
strand direction is reversed).

Picard and sambamba pick up these reads as duplicates

This change fixes this issue by checking the direction when doing the
left and right group bys in mark duplicates.